power_for_nA Function to calculate power for given number of after samples for a paire t-test of the difference in means before vs. after change using power.t.test

Function to calculate power for given number of after samples for a paired t-test of the difference in means before vs. after change using power.t.test

Usage

power_for_nA(
  nA,
  S,
  nB,
  delta,
  sd_within,
  sd_delta,
  alpha,
  distribution = c("normal", "nbinom", "binomial"),
  useTest = c("paired-t", "wilcoxon", "prop.test")
)

power_for_nA(
  nA,
  S,
  nB,
  delta,
  sd_within,
  sd_delta,
  alpha,
  distribution = c("normal", "nbinom", "binomial"),
  useTest = c("paired-t", "wilcoxon", "prop.test")
)

Arguments

nA

Number of after samples per site

S

Number of sites

nB

Number of before samples per site

delta

Hypothesized mean change

sd_within

Standard deviation within sites

sd_delta

Standard deviation of true changes among sites

alpha

Significance level

distribution

Distribution for simulated data. Must be "normal" here.

useTest

Which test to use. Must be "paired-t" here.

Value

Power for the given number of after sample

Power for the given number of after sample